Implement constant references in compiler IL (initially disabled)
Please refer to the corresponding OMR commit for the highest-level
explanation of what these changes are for. This commit implements the
project-specific functionality mentioned there and makes adjustments
elsewhere in the JIT to accommodate TR_ALLOW_NON_CONST_KNOWN_OBJECTS and
take advantage of const refs where possible.

Const refs are disabled in builds without J9VM_OPT_OPENJDK_METHODHANDLE
because they're incompatible with the legacy J9 MethodHandles for GC
reasons. In particular, the GC would essentially need to be aware of
custom thunks. With OpenJDK MethodHandles, customization happens at the
bytecode level, and the GC is already aware of the generated class (just
by virtue of its being a class). Note that const refs are currently
disabled by default in OMR anyway, but that is expected to change.

Keepalives are determined based on the compilation's root retained
method set and also its "global" keepalive classes (see
J9::Compilation::addKeepaliveClass()). An effort is made to detect and
ignore redundant keepalives, which are ones that are already covered by
a bond, an existing const ref, or another keepalive that has already
been chosen to be promoted to a const ref. With this, RetainedMethodSet
together with const refs truly will prevent JIT bodies from outliving
inlined code, with no keepalive caveat.

The const refs are sorted (and so grouped) by owning class, where the
owning class is assigned based on reachability in the const provenance
graph. This ordering minimizes the amount of GC bookkeeping that's
necessary for a given choice of owning classes.

At the end of the compilation, just after CH table commit, the const ref
slots are described to the GC and then initialized to point at the
correct objects. For each owning class, an instance of J9ConstRefArray
is added to the constRefArrays list of the J9Class. See the commit that
introduced J9ConstRefArray for a detailed description of the bookkeeping
and initialization process.

When code cache reclamation reclaims a JIT body, all bookkeeping for
const refs in that JIT body is immediately removed.

For JITServer, one additional round trip is required in each compilation
in which const refs reach code generation or in which there appear to be
keepalives. If there are (potentially redundant) keepalives, then the
round trip happens at the time of keepalive determination. The client
does its repeat retained methods analysis right away and remembers the
result, which is used to identify redundant keepalives in the same way
as for local compilations. Once the final keepalives have been
determined, they are communicated to the server along with all of the
const provenance edges that the client has collected throughout the
compilation. If OTOH there are definitely no keepalives, but there are
const refs, then the server will send a message to the client purely in
order to get the const provenance edges. Either way, the server will
then have the complete const provenance graph, which it will use to
assign owning classes to the const refs as usual. At the very end of the
compilation, the const refs are described to the GC and initialized on
the client using some additional information provided with the server's
compilationCode message.