Generate inline superclass test for Class.isAssignableFrom on Z
Previously when performing Class.isAssignableFrom, only the trivial case
of class equality was done inline and any other case relied on a callout
to a C Helper.

With this change, an inlined superclass test is added along with a
castClassCache test, avoiding the C Helper in the most common cases.

Additionally, this change checks for class depth at compile time when
possible to fast-fail when fromClassDepth <=  toClassDepth.

Signed-off-by: Matthew Hall <matthew.hall3@outlook.com>